+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Record of the modification in project development.
 - [Issue 5 : OVERWRITEINSTALL confirmation is missing](https://github.com/suikan4github/kaiten-yaki/issues/5)
 - [Issue 6 : Remove loglevel dependency from the void-kaiten-yaki.sh ](https://github.com/suikan4github/kaiten-yaki/6)
 - [Issue 7 : Add the return status validation ](https://github.com/suikan4github/kaiten-yaki/7)
+- [Issue 11 : Make chroot'ed job independent script file ](https://github.com/suikan4github/kaiten-yaki/11)
 
 ### Deprecated
 ### Removed